+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [release-80] - 2022-06-16
+
+- RMI-341: Created mechanism to measure customer effort score
+
 ## [release-79] - 2022-03-24
 
 - RMI-504: Dependency updates, to fix security vulnerabilities. (Manual Update)
@@ -420,6 +424,8 @@
 
 Initial release
 
+[release-80]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-79...release-80
+[release-79]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-78...release-79
 [release-78]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-77...release-78
 [release-77]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-76...release-77
 [release-76]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-75...release-76